Refactoring after mentoring
