+++ CampIoasys2Fase/ViewModel/LoginViewModel.swift
@@ -8,11 +8,19 @@
 import Foundation
 
 class LoginViewModel {
+    
+    // MARK: Atributes
     var api = API()
     var network = Networking()
     
-    func doLogin(parameters : [String : Any]?, headers : [String : String] , completion : @escaping (Any?, String?) -> ()){
+    let headers = ["Content-Type" : "application/json"]
+    
+    // MARK: Methods
+    
+    func doLogin(parameters : [String : Any]?, completion : @escaping (Any?, String?) -> ()){
         network.performRequest(type: LoginResult.self, path: api.returnEndpoint(endpoint: .login), method: .post, headers : headers, parameters: parameters) { [self] result, error in
+            
+            
             if let error = error {
                 completion("Deu erro", error.localizedDescription)
             }
